+++ .env.production
@@ -1,2 +1,3 @@
 REACT_APP_API='https://api.icetea.io'
+REACT_APP_RPC='https://rpc.icetea.io'
 REACT_APP_INIT_BLOCK=1
