+++ src/service/tweb3/index.js
@@ -1,3 +1,4 @@
-export * from './blocks'
-export * from './transactions'
-export * from './contracts'
+export * from './blocks';
+export * from './transactions';
+export * from './contracts';
+export { default as tweb3 } from './tweb3';
