+++ src/service/tweb3/transactions.js
@@ -1,4 +1,4 @@
-import tweb3 from '../../tweb3';
+import tweb3 from './tweb3';
 // import store from '../../store';
 // import * as actions from '../../store/actions';
